Merge from master

+++ src/main/java/il/ac/afeka/cloud/enums/TimeEnum.java
@@ -3,7 +3,7 @@ package il.ac.afeka.cloud.enums;
 public enum TimeEnum {
 	lastDay(1000 * 60 * 60 * 24),
 	lastWeek(1000 * 60 * 60 * 24 * 7),
-	lastMonth(1000 * 60 * 60 * 24 * 30);
+	lastMonth((long)(1000) * 60 * 60 * 24 * 30);
 	
 	private final long millisecs;
